Rename files, add canUndo/canRedo functions
