+++ src/main/js/models/helpers/changeCell.js
@@ -13,7 +13,7 @@ import {
 import request from "superagent";
 import * as f from "lodash/fp";
 import Raven from "raven-js";
-import {remember} from "../../components/table/undo/undoer";
+import {remember} from "../../components/table/undo/tableHistory";
 
 async function changeCell({cell, value, options = {}}) {
   window.devLog(`Changing ${cell.kind} cell ${cell.id} from`, cell.value, "to", (value.value || value));
